move recent properties out of css editor
